Fail generating website if any page throws an error
This will prevent us from having pages on the site that just show error
stack traces -- instead, Travis will fail with an error and we'll
notice sooner.

I cleaned up this logic to use promises which was eay easier to track
for the error behavior and should be simpler to follow regardless.